pyproject.toml
[project]
name = "llm-verbal-skills-trainer"
version = "0.1.0"
description = "LLM model which helps candidate to enhance their verbal skills"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "gradio (>=5.20.0,<6.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "llvmlite (>=0.44.0,<0.45.0)",
    "numba (>=0.61.0,<0.62.0)",
    "openai-whisper (>=20240930,<20240931)",
    "protobuf (>=5.29.3,<6.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "accelerate (>=1.4.0,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "noisereduce (>=3.0.3,<4.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "plotly (>=6.0.0,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"



